Add a cast to avoid MSVC warning
This fixes the following warning:
../src/decode.c(666): warning C4244: '=': conversion from 'uint16_t' to 'uint8_t', possible loss of data

+++ src/decode.c
@@ -663,7 +663,7 @@ static void read_vartx_tree(Dav1dTileContext *const t,
         b->uvtx = dav1d_max_txfm_size_for_bs[bs][f->cur.p.layout];
     }
     assert(!(tx_split[0] & ~0x33));
-    b->tx_split0 = tx_split[0];
+    b->tx_split0 = (uint8_t)tx_split[0];
     b->tx_split1 = tx_split[1];
 }
